+++ changelog.md
@@ -1,6 +1,7 @@
 2.0.1 - xxxx-xx-xx
 --
 * Upgrade bootstrap v4 from alpha 2 to alpha 3.
+* Adjust form submissions filter to search with ActiveRelation#or.
 
 
 2.0.0 - 2016-07-29